+++ Graph Theory/kosaraju.cpp
@@ -16,7 +16,7 @@ void dfs1(int u){
 	seen[u] = true;
 
 	// For every vertex v adjacent to u.
-	for (int i = 0; i < (int)g[u].size(); i++){
+	for (int i = 0; i < g[u].size(); i++){
 		int v = g[u][i];
 
 		// If vertex v wasn't already visited.
@@ -36,7 +36,7 @@ void dfs2(int u){
 	component[nc].push_back(u);
 
 	// For every vertex v adjacent to u.
-	for (int i = 0; i < (int)gt[u].size(); i++){
+	for (int i = 0; i < gt[u].size(); i++){
 		int v = gt[u][i];
 
 		// If vertex v wasn't already visited.
