+++ Graph Theory/prim.cpp
@@ -48,7 +48,7 @@ void prim(int u){
 			in_mst[u] = true;
 
 			// For every vertex v adjacent to u.
-			for (int i = 0; i < (int)g[u].size(); i++){
+			for (int i = 0; i < g[u].size(); i++){
 				int v = g[u][i].first;
 				int w = g[u][i].second;
