+++ Number Theory/euler_phi_linear_sieve.cpp
@@ -16,7 +16,7 @@ vector<int> linear_sieve(int n){
 		}
 
 		// Since every number x can be written as x = i * mp[x] we can iterate over every prime mp[x] (prime[j]).
-		for (int j = 0; j < (int)p.size() and p[j] <= mp[i] and i * p[j] <= n; j++){
+		for (int j = 0; j < p.size() and p[j] <= mp[i] and i * p[j] <= n; j++){
 			// phi[x] = phi[i * mp[x]] = phi[i] * mp[x] if mp[x] divides i or phi[i] * (mp[x] - 1) if mp[x] doesn't divide i.
 			phi[i * p[j]] = phi[i] * (p[j] == mp[i] ? p[j] : p[j] - 1);
 			mp[i * p[j]] = p[j];
