+++ Number Theory/n_choose_k_modulo_m.cpp
@@ -179,14 +179,14 @@ void chinese_remainder_theorem(vector<long long> a, vector<long long> m, long lo
 	long long a2, m2, x1, y1, gcd, lcm;
 
 	// Making 0 <= ai < mi.
-	for (int i = 0; i < (int)a.size(); i++){
+	for (int i = 0; i < a.size(); i++){
 		a[i] = ((a[i] % m[i]) + m[i]) % m[i];
 	}
 
 	a1 = a[0];
 	m1 = m[0];
 
-	for (int i = 1; i < (int)a.size(); i++){
+	for (int i = 1; i < a.size(); i++){
 		a2 = a[i];
 		m2 = m[i];
 
@@ -259,7 +259,7 @@ long long nck(long long n, long long k, long long m){
 	}
 
 	// O(Log(N) * Log(M)) - Calculating n choose k modulo each prime power that composes m.
-	for (int i = 0; i < (int)factor.size(); i++){
+	for (int i = 0; i < factor.size(); i++){
 		p = factor[i].first;
 		e = factor[i].second;
 		pe = fast_exp(p, e, m + 1);
@@ -296,7 +296,7 @@ void nck_init(long long m){
 	dp_f.resize(factor.size());
 
 	// O(M) - For each prime power that composes m.
-	for (int i = 0; i < (int)factor.size(); i++){
+	for (int i = 0; i < factor.size(); i++){
 		p = factor[i].first;
 		e = factor[i].second;
 		pe = fast_exp(p, e, m + 1);
