+++ String Algorithms/z.cpp
@@ -5,17 +5,17 @@ vector<int> z_function(const string &s){
 
 	// Initialization.
 	z.resize(s.size());
-	z[0] = (int)s.size(); // We can ignore this value.
+	z[0] = s.size(); // We can ignore this value.
 	int l = 0, r = 0; // Range is [L, R)
 
-	for (int i = 1; i < (int)s.size(); i++){
+	for (int i = 1; i < s.size(); i++){
 		if (i > r){
 			l = r = i;
 		}
 
 		z[i] = min(z[i - l], r - i);
 
-		while (i + z[i] < (int)s.size() and s[i + z[i]] == s[z[i]]){
+		while (i + z[i] < s.size() and s[i + z[i]] == s[z[i]]){
 			z[i]++;
 			l = i;
 			r = i + z[i];
